refactor(worktree): consolidate validate_branch to crate::agent_ops (Task #9 Option C epilogue)
PR #18 consolidated 7 shared agent-level helpers between `src/ops.rs`
and `src/mcp/handlers.rs` into the canonical `src/agent_ops.rs` module.
This epilogue closes out a 3rd byte-identical duplicate of
`validate_branch` that was living privately inside `src/worktree.rs`.

Pre-edit verification: the `worktree.rs` private copy (former lines
47–54) was byte-for-byte identical to `crate::agent_ops::validate_branch`
(only the leading `pub` visibility + doc-comment wrapping differed).

Changes:
- `src/worktree.rs`: add `use crate::agent_ops::validate_branch;`,
  delete the inline `fn validate_branch` duplicate (and its doc comment).
- `src/agent_ops.rs::tests`: migrate the two tests
  (`test_validate_branch_valid`, `test_validate_branch_rejects`) that
  lived in `worktree.rs::tests` — colocated with their subject now that
  `validate_branch` canonically lives in `agent_ops`.

CI gates:
- `cargo fmt --check` clean
- `cargo clippy --all-targets -- -D warnings` clean
- `cargo test --release` — 623 tests pass (unchanged)
