+++ src/agent_ops.rs
@@ -293,6 +293,27 @@ mod tests {
         assert!(!validate_branch("main\ninjected"));
     }
 
+    // Migrated from `src/worktree.rs::tests` as part of Task #9 Option C
+    // epilogue (worktree.rs no longer holds its own `validate_branch` copy).
+
+    #[test]
+    fn test_validate_branch_valid() {
+        assert!(validate_branch("main"));
+        assert!(validate_branch("feature/my-branch"));
+        assert!(validate_branch("agend/agent-1"));
+        assert!(validate_branch("v1.0.0"));
+    }
+
+    #[test]
+    fn test_validate_branch_rejects() {
+        assert!(!validate_branch(""));
+        assert!(!validate_branch(".."));
+        assert!(!validate_branch("foo/../bar"));
+        assert!(!validate_branch("-starts-with-dash"));
+        assert!(!validate_branch("has spaces"));
+        assert!(!validate_branch("has;semicolon"));
+    }
+
     // --- merge_metadata (2 from ops.rs) ---
 
     #[test]
